attendance_app/attendance_backend/requirements-sync.txt
fastapi
psycopg2-binary
google-cloud-firestore
pydantic
